requirements.txt
amqp==5.1.1
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
celery==5.3.4
ConfigParser==6.0.0
contextlib2==21.6.0
cryptography==41.0.4
Cython==3.0.2
dl==0.1.0
docutils==0.20.1
googletrans==4.0.0rc1
HTMLParser==0.0.2
ipython==8.15.0
ipywidgets==8.1.1
Jinja2==3.1.2
keyring==24.2.0
lockfile==0.12.2
mock==5.1.0
numpy==1.26.0
python-jose==3.3.0
passlib==1.7.4
Pillow==10.0.1
protobuf==4.24.3
pydantic==2.3.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
railroad==0.5.0
simplejson==3.19.1
Sphinx==7.2.6
SQLAlchemy==2.0.21
tornado==6.3.3
trove_classifiers==2023.9.19
truststore==0.8.0
urllib3_secure_extra==0.1.0
pypi-xmlrpc==2020.12.3
zstandard==0.21.0
lxml==4.9.3
qrcode==7.4.2
python-multipart==0.0.6
psycopg2-binary==2.9.7
bcrypt==4.0.1
pyjnius==1.5.0
fastapi==0.103.1
uvicorn==0.23.2
xmltodict==0.13.0
